Move Msf::Java to the normal Msf::Exploit::Remote namespace
