+++ lib/msf/core.rb
@@ -75,10 +75,6 @@ require 'msf/http/jboss'
 # Kerberos Support
 require 'msf/kerberos/client'
 
-# Java RMI Support
-require 'msf/java/rmi/util'
-require 'msf/java/rmi/client'
-
 # Drivers
 require 'msf/core/exploit_driver'
