+++ lib/msf/core/exploit/java.rb
@@ -14,6 +14,8 @@
 ###
 
 require 'msf/core'
+require 'msf/core/exploit/java/rmi/util'
+require 'msf/core/exploit/java/rmi/client'
 
 module Msf
 module Exploit::Java
