+++ modules/exploits/multi/misc/java_rmi_server.rb
@@ -8,7 +8,7 @@ require 'msf/core'
 class Metasploit3 < Msf::Exploit::Remote
   Rank = ExcellentRanking
 
-  include Msf::Java::Rmi::Client
+  include Msf::Exploit::Remote::Java::Rmi::Client
   include Msf::Exploit::Remote::HttpServer
 
   def initialize(info = {})
